Modularity and new feature done
