+++ ColorCoding.h
@@ -1,0 +1,6 @@
+#ifndef COLOR_CODING_H
+#define COLOR_CODING_H
+
+void colorCodingManual();
+
+#endif
