+++ Tests.h
@@ -1,0 +1,15 @@
+#ifndef TESTS_H
+#define TESTS_H
+#include "ColorPair.h"
+#include "assert.h"
+
+void testNumberToPair(int pairNumber,
+    TelCoColorCoder::MajorColor expectedMajor,
+    TelCoColorCoder::MinorColor expectedMinor);
+
+void testPairToNumber(
+    TelCoColorCoder::MajorColor major,
+    TelCoColorCoder::MinorColor minor,
+    int expectedPairNumber);
+
+#endif
